+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   - Remove the special casing of small galaxies from orbital and defence techs that meant they almost never got dealt
   - Sudden Death Tech chance is now tied to galaxy size and distance travelled
 - Updated Protocol Kill-switch for addition of Helios death weapon
+- Updated Pelican Upgrade Tech to allow any unit to fire while being transported
 
 ### Bugfix